Uplift factor is parameter SKYE_UPLIFT. Like this it can always be pointing upwards (independent on control mode)

+++ src/ui/uas/UASSkyeControlAdvancedWidget.h
@@ -27,8 +27,8 @@ public:
     ~UASSkyeControlAdvancedWidget();
 
 signals:
-    void linearSliderValueChanged(float);
-    void angularSliderValueChanged(float);
+    void linearSliderValueChanged(double);
+    void angularSliderValueChanged(double);
     void liftSliderValueChanged(double);
     void rollSliderValueChanged(double);
     void pitchSliderValueChanged(double);
